issue-996: extract memory contract/runtime
