+++ crates/tau-memory/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "tau-memory"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+tokio = { version = "1.43", features = ["time"] }
+
+# Internal crates
+
+tau-core = { path = "../tau-core" }
+tau-runtime = { path = "../tau-runtime" }
+tau-session = { path = "../tau-session" }
+
+[dev-dependencies]
+tempfile = "3.10"
+tokio = { version = "1.43", features = ["rt-multi-thread", "macros", "time"] }
