+++ crates/tau-memory/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod memory_contract;
+pub mod memory_runtime;
