+++ crates/tau-memory/testdata/memory-contract/invalid-duplicate-case-id.json
@@ -1,0 +1,77 @@
+{
+  "schema_version": 1,
+  "name": "invalid-duplicate-case-id",
+  "cases": [
+    {
+      "schema_version": 1,
+      "case_id": "duplicate-case",
+      "mode": "extract",
+      "scope": {
+        "workspace_id": "tau-core"
+      },
+      "input_text": "Remember release cadence.",
+      "expected": {
+        "outcome": "success",
+        "entries": [
+          {
+            "memory_id": "mem-duplicate-case",
+            "summary": "Remember release cadence.",
+            "tags": [
+              "remember"
+            ],
+            "facts": [
+              "scope=tau-core"
+            ],
+            "source_event_key": "tau-core:extract:duplicate-case",
+            "recency_weight_bps": 9000,
+            "confidence_bps": 8200
+          }
+        ]
+      }
+    },
+    {
+      "schema_version": 1,
+      "case_id": "duplicate-case",
+      "mode": "retrieve",
+      "scope": {
+        "workspace_id": "tau-core"
+      },
+      "query_text": "release",
+      "prior_entries": [
+        {
+          "memory_id": "mem-release",
+          "summary": "Release checklist includes rollback and canary verification.",
+          "tags": [
+            "release",
+            "rollback"
+          ],
+          "facts": [
+            "owner=sre"
+          ],
+          "source_event_key": "tau-core:extract:release",
+          "recency_weight_bps": 9400,
+          "confidence_bps": 9000
+        }
+      ],
+      "expected": {
+        "outcome": "success",
+        "entries": [
+          {
+            "memory_id": "mem-release",
+            "summary": "Release checklist includes rollback and canary verification.",
+            "tags": [
+              "release",
+              "rollback"
+            ],
+            "facts": [
+              "owner=sre"
+            ],
+            "source_event_key": "tau-core:extract:release",
+            "recency_weight_bps": 9400,
+            "confidence_bps": 9000
+          }
+        ]
+      }
+    }
+  ]
+}
