+++ crates/tau-memory/testdata/memory-contract/retrieve-ranking.json
@@ -1,0 +1,96 @@
+{
+  "schema_version": 1,
+  "name": "retrieve-ranking",
+  "description": "Valid retrieval fixture for deterministic ranking and compatibility replay.",
+  "cases": [
+    {
+      "schema_version": 1,
+      "case_id": "retrieve-release-rollback-priority",
+      "mode": "retrieve",
+      "scope": {
+        "workspace_id": "tau-core",
+        "channel_id": "telegram:ops",
+        "actor_id": "user-84"
+      },
+      "query_text": "rollback release",
+      "retrieval_limit": 2,
+      "prior_entries": [
+        {
+          "memory_id": "mem-release-checklist",
+          "summary": "Release checklist includes rollback and canary verification.",
+          "tags": [
+            "release",
+            "rollback"
+          ],
+          "facts": [
+            "owner=sre"
+          ],
+          "source_event_key": "tau-core:extract:release-checklist",
+          "recency_weight_bps": 9400,
+          "confidence_bps": 9000
+        },
+        {
+          "memory_id": "mem-style-preference",
+          "summary": "User prefers concise pull request descriptions.",
+          "tags": [
+            "style",
+            "docs"
+          ],
+          "facts": [
+            "owner=devexp"
+          ],
+          "source_event_key": "tau-core:extract:style-preference",
+          "recency_weight_bps": 9200,
+          "confidence_bps": 8800
+        },
+        {
+          "memory_id": "mem-incident-rollback-drill",
+          "summary": "Incident drills require rollback rehearsals every sprint.",
+          "tags": [
+            "incident",
+            "rollback"
+          ],
+          "facts": [
+            "owner=sre"
+          ],
+          "source_event_key": "tau-core:extract:incident-rollback-drill",
+          "recency_weight_bps": 9600,
+          "confidence_bps": 9100
+        }
+      ],
+      "expected": {
+        "outcome": "success",
+        "entries": [
+          {
+            "memory_id": "mem-release-checklist",
+            "summary": "Release checklist includes rollback and canary verification.",
+            "tags": [
+              "release",
+              "rollback"
+            ],
+            "facts": [
+              "owner=sre"
+            ],
+            "source_event_key": "tau-core:extract:release-checklist",
+            "recency_weight_bps": 9400,
+            "confidence_bps": 9000
+          },
+          {
+            "memory_id": "mem-incident-rollback-drill",
+            "summary": "Incident drills require rollback rehearsals every sprint.",
+            "tags": [
+              "incident",
+              "rollback"
+            ],
+            "facts": [
+              "owner=sre"
+            ],
+            "source_event_key": "tau-core:extract:incident-rollback-drill",
+            "recency_weight_bps": 9600,
+            "confidence_bps": 9100
+          }
+        ]
+      }
+    }
+  ]
+}
